// frontend/src/core/models/accountsCstm.ts
/**
 * Created by @ES Express Systems
 * User: Rafael Gutierrez Gaspar
 * Date: Mon Jan 25 2021 02:40:50 GMT-0400 (Bolivia Time)
 * Time: 2:40:50
 * Last User updated: Rafael Gutierrez Gaspar
 * Last date updated: Mon Jan 25 2021 02:40:50 GMT-0400 (Bolivia Time)
 * Last time updated: 2:40:50
 *
 * Caution: es-sections will be replaced by script execution
 */

//<es-section>

//</es-section>

export class AccountsCstm {

  //<es-section>
  
  id_c: string;
  
  
  
  
  
  jjwg_maps_geocode_status_c: string;
  
  jjwg_maps_address_c: string;
  
  numero_documento_c: string;
  
  extension_documento_c: string;
  
  
  
  
  
  
  
  jjwg_maps_lng_c: number;
  
  jjwg_maps_lat_c: number;
  
  
  
  
  
  

  //</es-section>
}