// frontend/src/core/models/accountsContacts.ts
/**
 * Created by @ES Express Systems
 * User: Rafael Gutierrez Gaspar
 * Date: Mon Jan 25 2021 02:40:49 GMT-0400 (Bolivia Time)
 * Time: 2:40:49
 * Last User updated: Rafael Gutierrez Gaspar
 * Last date updated: Mon Jan 25 2021 02:40:49 GMT-0400 (Bolivia Time)
 * Last time updated: 2:40:49
 *
 * Caution: es-sections will be replaced by script execution
 */
import {Contacts} from "./contacts";
import {Accounts} from "./accounts";

//<es-section>

//</es-section>

export class AccountsContacts {

  //<es-section>
  
  id: string;
  
  
  
  
  
  contact_id: string;
  
  account_id: string;
  
  
  
  
  date_modified: Date;
  
  
  
  
  
  
  
  deleted: number;
  
  
  

  //</es-section>
  
  accountContactContacts:Contacts;
  
  accountContactAccounts:Accounts;
}