// frontend/src/core/models/aodIndexevent.ts
/**
 * Created by @ES Express Systems
 * User: Rafael Gutierrez Gaspar
 * Date: Mon Jan 25 2021 02:41:11 GMT-0400 (Bolivia Time)
 * Time: 2:41:11
 * Last User updated: Rafael Gutierrez Gaspar
 * Last date updated: Mon Jan 25 2021 02:41:11 GMT-0400 (Bolivia Time)
 * Last time updated: 2:41:11
 *
 * Caution: es-sections will be replaced by script execution
 */

//<es-section>

//</es-section>

export class AodIndexevent {

  //<es-section>
  
  id: string;
  
  
  
  
  
  name: string;
  
  error: string;
  
  record_module: string;
  
  
  description: string;
  
  
  
  date_entered: Date;
  
  date_modified: Date;
  
  
  
  modified_user_id: string;
  
  created_by: string;
  
  assigned_user_id: string;
  
  record_id: string;
  
  
  
  
  
  deleted: number;
  
  success: number;
  
  
  

  //</es-section>
}