// frontend/src/core/models/accountsOpportunities.ts
/**
 * Created by @ES Express Systems
 * User: Rafael Gutierrez Gaspar
 * Date: Mon Jan 25 2021 02:40:51 GMT-0400 (Bolivia Time)
 * Time: 2:40:51
 * Last User updated: Rafael Gutierrez Gaspar
 * Last date updated: Mon Jan 25 2021 02:40:51 GMT-0400 (Bolivia Time)
 * Last time updated: 2:40:51
 *
 * Caution: es-sections will be replaced by script execution
 */
import {Opportunities} from "./opportunities";
import {Accounts} from "./accounts";

//<es-section>

//</es-section>

export class AccountsOpportunities {

  //<es-section>
  
  id: string;
  
  
  
  
  
  opportunity_id: string;
  
  account_id: string;
  
  
  
  
  date_modified: Date;
  
  
  
  
  
  
  
  deleted: number;
  
  
  

  //</es-section>
  
  accountOpportunityOpportunities:Opportunities;
  
  accountOpportunityAccounts:Accounts;
}